requirements.txt
pybit>=5.7.0
pandas>=2.2.0
numpy>=1.26.0
statsmodels>=0.14.0
scipy>=1.12.0
python-dotenv>=1.0.0
tabulate>=0.9.0
requests>=2.31.0
